Un-nest internal files Improve to test to check that state is correctly passed to all IChangeToken Some cleaning (license header, formatting, etc...)
