+++ test/Microsoft.AspNet.FileProviders.Combined.Tests/project.json
@@ -8,7 +8,6 @@
     "commands": {
         "test": "xunit.runner.aspnet"
     },
-    "resource": "File.txt;sub/**;Resources/**",
     "frameworks": {
         "dnx451": { },
         "dnxcore50": {
